Remove unused ticker module, add ItemsNew json config
Ticker module was removed due to not being used, and to clean up the codebase. The functionality provided by the ticker (scheduling tasks at certain times) has been replaced by more efficient and flexible solutions
that are better integrated into the rest of our code.
The json config "ItemsNew.json" was added to the "config/jsons" path to meet the need of the new feature's requirement.
This new config file introduces more granular control over various aspects of the game items such as their unlocking mechanism, mode of production, price etc.
This file will be used heavily to make tweak, balance and scaling of the game easier.

+++ dao/redis/player.go
@@ -19,7 +19,7 @@ func NewPlayerInfo() *PlayerInfo {
 }
 
 // 读取玩家数据
-func (r *Module) GetPlayerInfo(userID uint32, initPlayerData *pb.PBPlayerData) error {
+func (r *Module) GetPlayerInfo(userID int64, initPlayerData *pb.PBPlayerData) error {
 	bytes, _ := r.Client.Get(r.Ctx, r.player.TableName+strconv.Itoa(int(userID))).Bytes()
 	if bytes != nil {
 		err := json.Unmarshal(bytes, initPlayerData)
@@ -30,7 +30,7 @@ func (r *Module) GetPlayerInfo(userID uint32, initPlayerData *pb.PBPlayerData) e
 }
 
 // 保存玩家数据
-func (r *Module) SetPlayerInfo(userID uint32, playerInfo *pb.PBPlayerData) error {
+func (r *Module) SetPlayerInfo(userID int64, playerInfo *pb.PBPlayerData) error {
 	marshal, err := json.Marshal(playerInfo)
 	if err != nil {
 		return err
